refactor(mcp): split the tool_router impl into per-domain tools/ modules
The 16-tool #[tool_router] impl CartogServer moves out of lib.rs into
tools/{index,graph,search,rag,manage}.rs, each a #[tool_router(router=...)]
block. lib.rs keeps the 8 non-tool methods in a plain impl and hand-writes
tool_router() = index_router() + graph_router() + ... so Self::tool_router()
(field init, #[tool_handler] default, tests) resolves unchanged. lib.rs drops
2085 -> ~1170 lines.

tools/ is a child module of lib.rs, so it reaches lib.rs's private items
directly; no visibility widening was needed (lib.rs keeps its baseline
private surface). Test modules import EdgeKind/MAX_SEARCH_LIMIT/Parameters
from their origin crates. All 16 tools verified present in list_all();
metadata and bodies byte-identical; 114 tests unchanged.
