+++ crates/cartog-mcp/src/lib.rs
@@ -12,25 +12,24 @@ use std::sync::{Arc, Mutex};
 
 use rmcp::schemars;
 use rmcp::{
-    handler::server::{router::tool::ToolRouter, tool::schema_for_output, wrapper::Parameters},
+    handler::server::{router::tool::ToolRouter, tool::schema_for_output},
     model::*,
-    service::RequestContext,
-    tool, tool_handler, tool_router,
+    tool_handler,
     transport::stdio,
-    ErrorData as McpError, RoleServer, ServerHandler, ServiceExt,
+    ErrorData as McpError, ServerHandler, ServiceExt,
 };
 use schemars::JsonSchema;
 use serde::Serialize;
 use tracing::{debug, info};
 
-use cartog_core::{Compact, EdgeKind};
-use cartog_db::{Database, PinnedAttach, MAX_SEARCH_LIMIT};
+use cartog_db::{Database, PinnedAttach};
 use cartog_indexer as indexer;
 use cartog_rag as rag;
 use cartog_watch as watch;
 use cartog_watch::{StaleSnapshot, WatchConfig, WatchHandle};
 
 mod progress;
+mod tools;
 mod types;
 
 use types::*;
@@ -917,7 +916,6 @@ impl AtomicRole {
     }
 }
 
-#[tool_router]
 impl CartogServer {
     /// Construct a writable primary MCP server.
     ///
@@ -1124,931 +1122,16 @@ impl CartogServer {
         }
     }
 
-    /// Build or rebuild the code graph index for a directory.
-    #[tool(
-        description = "Build or rebuild the code graph index. Run this first before any other cartog tool, or after making code changes to keep the graph current. Incremental by default — only re-indexes changed files. Use force=true if results seem stale. Not for: routine queries (call once per session, not before every read). Returns: {files_indexed, files_skipped, symbols_added, edges_added, edges_resolved, edges_lsp_resolved, edges_marked_unresolvable, edges_marked_external}.",
-        annotations(
-            title = "Index codebase",
-            read_only_hint = false,
-            destructive_hint = false,
-            idempotent_hint = false,
-            open_world_hint = false
-        )
-    )]
-    async fn cartog_index(
-        &self,
-        Parameters(params): Parameters<IndexParams>,
-        ctx: RequestContext<RoleServer>,
-    ) -> Result<CallToolResult, McpError> {
-        if let Some(err) = self.refuse_if_read_only("cartog_index") {
-            return Err(err);
-        }
-        let path = params.path;
-        let force = params.force;
-        let db = Arc::clone(&self.db);
-        let cwd = Arc::clone(&self.cwd);
-        let redact = self.redact;
-        let walk_filter = Arc::clone(&self.walk_filter);
-        #[cfg(feature = "lsp")]
-        let lsp_manager = Arc::clone(&self.lsp_manager);
-        #[cfg(not(feature = "lsp"))]
-        let lsp_manager: () = ();
-        #[cfg(feature = "lsp")]
-        let lsp_unavailable = Arc::clone(&self.lsp_unavailable);
-        #[cfg(not(feature = "lsp"))]
-        let lsp_unavailable: () = ();
-
-        let (progress_tx, forwarder) = match ctx.meta.get_progress_token() {
-            Some(token) => {
-                let notifier = progress::peer_notifier(ctx.peer.clone());
-                let fwd = progress::spawn_forwarder(token, notifier);
-                (Some(fwd.tx.clone()), Some(fwd))
-            }
-            None => (None, None),
-        };
-
-        let cancel = ctx.ct.clone();
-        let join = tokio::task::spawn_blocking(move || {
-            let validated = validate_path_within_cwd_canonical(&path, &cwd).map_err(mcp_err)?;
-            debug!(path = %validated.display(), force, "indexing directory");
-            let probe = || cancel.is_cancelled();
-            let probe_ref: Option<&(dyn Fn() -> bool + Send + Sync)> = Some(&probe);
-            let result = index_with_optional_lsp(
-                &db,
-                &lsp_manager,
-                &validated,
-                force,
-                progress_tx.clone(),
-                probe_ref,
-                redact,
-                walk_filter.as_ref(),
-            )?;
-            let result = catch_up_lsp(
-                &db,
-                &lsp_manager,
-                &lsp_unavailable,
-                &validated,
-                progress_tx,
-                probe_ref,
-                result,
-            )?;
-
-            let json = serde_json::to_string_pretty(&result)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            // Human summary first, then the raw JSON, so the agent sees counts at a glance.
-            let mut text = format!("{}\n{json}", indexer::render_index_summary(&result));
-            if let Some(hint) = suggestions_for("cartog_index") {
-                text.push_str("\n\n");
-                text.push_str(hint);
-            }
-            Ok(CallToolResult::success(vec![Content::text(text)]))
-        })
-        .await;
-
-        // Drain the forwarder unconditionally — even on join error or tool
-        // error — so the spawned task doesn't leak waiting on `rx.recv()`.
-        // The blocking closure already dropped its progress_tx clone; the
-        // Forwarder still holds one, so we drop it here to close the channel.
-        if let Some(fwd) = forwarder {
-            drop(fwd.tx);
-            let _ = fwd.join.await;
-        }
-
-        join.map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Show symbols and structure of a file without reading its content.
-    #[tool(
-        description = "Show one file's structure: functions, classes, methods, imports with signatures and line ranges. Use this INSTEAD of reading a file when you need to understand what's in it — then Read only the specific lines you need. For understanding how a FEATURE or AREA works (spanning files), prefer cartog_context — it returns the relevant bodies across files in one call. Not for: reading the actual function body (use Read with offset/limit), or finding usages (use cartog_refs). Returns: Symbol[] with {name, kind, signature, line_start, line_end, parent_id, is_async, is_exported}.",
-        annotations(title = "Outline file", read_only_hint = true, open_world_hint = false),
-        output_schema = output_schema_for::<SymbolList>()
-    )]
-    async fn cartog_outline(
-        &self,
-        Parameters(params): Parameters<OutlineParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let file = params.file;
-        let db = Arc::clone(&self.db);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            debug!(file = %file, "outline");
-            let db = db.lock().map_err(|_| {
-                mcp_err("internal error: database lock poisoned (server restart required)")
-            })?;
-            let mut symbols = db
-                .outline(&file)
-                .map_err(|e| mcp_err(format!("outline query failed: {e}")))?;
-            if mcp_compact() {
-                symbols.compact_in_place();
-            }
-
-            let json = serde_json::to_string_pretty(&symbols)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(SymbolList { results: symbols }).ok();
-            tool_response(&db, json, structured, "cartog_outline", stale)
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Find all references to a symbol (calls, imports, inherits, type references, raises).
-    #[tool(
-        description = "Find all usages of a symbol across the codebase. Use when asked 'where is X used?', 'who calls X?', 'who imports X?'. Filter by kind: calls, imports, inherits, references, raises. Requires an exact symbol name — use cartog_search first if unsure of the name. Not for: discovering what a function calls (use cartog_callees), or transitive impact (use cartog_impact). Returns: array of {edge: {kind, target_name, line}, source: Symbol | null}.",
-        annotations(
-            title = "Find references",
-            read_only_hint = true,
-            open_world_hint = false
-        ),
-        output_schema = output_schema_for::<RefList>()
-    )]
-    async fn cartog_refs(
-        &self,
-        Parameters(params): Parameters<RefsParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let name = params.name;
-        let kind_str = params.kind;
-        let db = Arc::clone(&self.db);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            let kind_filter = kind_str
-                .as_deref()
-                .map(|s| {
-                    s.parse::<EdgeKind>().map_err(|_| {
-                        mcp_err(format!(
-                            "invalid edge kind '{s}'. \
-                             Valid: calls, imports, inherits, references, raises"
-                        ))
-                    })
-                })
-                .transpose()?;
-
-            debug!(name = %name, kind = ?kind_filter, "refs");
-            let db = db.lock().map_err(|_| {
-                mcp_err("internal error: database lock poisoned (server restart required)")
-            })?;
-            let results = db
-                .refs(&name, kind_filter)
-                .map_err(|e| mcp_err(format!("refs query failed: {e}")))?;
-
-            let compact = mcp_compact();
-            let entries: Vec<RefEntry> = results
-                .into_iter()
-                .map(|(edge, mut sym)| {
-                    if compact {
-                        if let Some(s) = sym.as_mut() {
-                            s.compact_in_place();
-                        }
-                    }
-                    RefEntry { edge, source: sym }
-                })
-                .collect();
-
-            let json = serde_json::to_string_pretty(&entries)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(RefList { results: entries }).ok();
-            tool_response_named(&db, json, structured, "cartog_refs", Some(&name), stale)
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Find what a symbol calls.
-    #[tool(
-        description = "Trace what a function calls. Use when asked 'what does X call?', 'show me the call graph of X', or to understand execution flow. Requires an exact symbol name. Not for: finding who calls a function (use cartog_refs with kind=calls). Returns: Edge[] of {kind, target_name, line, file}.",
-        annotations(
-            title = "Trace callees",
-            read_only_hint = true,
-            open_world_hint = false
-        ),
-        output_schema = output_schema_for::<EdgeList>()
-    )]
-    async fn cartog_callees(
-        &self,
-        Parameters(params): Parameters<CalleesParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let name = params.name;
-        let db = Arc::clone(&self.db);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            debug!(name = %name, "callees");
-            let db = db.lock().map_err(|_| {
-                mcp_err("internal error: database lock poisoned (server restart required)")
-            })?;
-            let edges = db
-                .callees(&name)
-                .map_err(|e| mcp_err(format!("callees query failed: {e}")))?;
-
-            let json = serde_json::to_string_pretty(&edges)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(EdgeList { results: edges }).ok();
-            tool_response_named(&db, json, structured, "cartog_callees", Some(&name), stale)
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Transitive impact analysis — what breaks if this symbol changes?
-    #[tool(
-        description = "Assess blast radius before refactoring. Shows everything that transitively depends on a symbol up to N hops. Use when asked 'what breaks if I change X?', 'is it safe to rename/delete X?', or before any rename/extract/move/delete refactoring. Not for: direct callers only (use cartog_refs), or what the symbol calls (use cartog_callees). Returns: array of {edge, depth} where depth=1 is direct, depth=2 is one hop away, etc.",
-        annotations(
-            title = "Impact analysis",
-            read_only_hint = true,
-            open_world_hint = false
-        ),
-        output_schema = output_schema_for::<ImpactList>()
-    )]
-    async fn cartog_impact(
-        &self,
-        Parameters(params): Parameters<ImpactParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let name = params.name;
-        let depth = params.depth.unwrap_or(3).min(MAX_IMPACT_DEPTH);
-        let db = Arc::clone(&self.db);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            debug!(name = %name, depth, "impact");
-            let db = db.lock().map_err(|_| {
-                mcp_err("internal error: database lock poisoned (server restart required)")
-            })?;
-            let results = db
-                .impact(&name, depth)
-                .map_err(|e| mcp_err(format!("impact query failed: {e}")))?;
-
-            let entries: Vec<ImpactEntry> = results
-                .into_iter()
-                .map(|(edge, d)| ImpactEntry { edge, depth: d })
-                .collect();
-
-            let json = serde_json::to_string_pretty(&entries)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(ImpactList { results: entries }).ok();
-            tool_response_named(&db, json, structured, "cartog_impact", Some(&name), stale)
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Find a call path between two symbols, with each hop's body inline.
-    #[tool(
-        description = "Trace how one symbol reaches another through the call graph. Returns the shortest call path from `from` to `to`, each hop carrying the calling symbol's body inline. Use when asked 'how does A reach B?', 'trace the execution flow from X to Y', 'what's the call path between X and Y?'. Only statically-resolved `calls` edges are followed (dynamic dispatch is not traced). Not for: all callers (use cartog_refs), or blast radius (use cartog_impact). Returns: {found: bool, hops: [{source_name, target_name, kind, file_path, line, body?}]}.",
-        annotations(
-            title = "Trace call path",
-            read_only_hint = true,
-            open_world_hint = false
-        ),
-        output_schema = output_schema_for::<TraceList>()
-    )]
-    async fn cartog_trace(
-        &self,
-        Parameters(params): Parameters<TraceParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let from = params.from;
-        let to = params.to;
-        let depth = params.depth.unwrap_or(8).min(MAX_TRACE_DEPTH);
-        let db = Arc::clone(&self.db);
-        let cwd = Arc::clone(&self.cwd);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            debug!(from = %from, to = %to, depth, "trace");
-            let db = db.lock().map_err(|_| {
-                mcp_err("internal error: database lock poisoned (server restart required)")
-            })?;
-            let path = db
-                .trace(&from, &to, depth)
-                .map_err(|e| mcp_err(format!("trace query failed: {e}")))?;
-
-            let compact = mcp_compact();
-            let hops: Vec<TraceHop> = path
-                .iter()
-                .flatten()
-                .map(|h| TraceHop {
-                    // Compact bounds each hop body to a snippet (preview), keeping
-                    // the path readable without shipping full function bodies.
-                    body: trace_hop_body(&db, &cwd, &h.source_id).map(|b| {
-                        if compact {
-                            rag::search::snippet(&b)
-                        } else {
-                            b
-                        }
-                    }),
-                    source_name: h.source_name.clone(),
-                    target_name: h.target_name.clone(),
-                    kind: h.kind.to_string(),
-                    file_path: h.file_path.clone(),
-                    line: h.line,
-                })
-                .collect();
-
-            let result = TraceList {
-                found: path.is_some(),
-                hops,
-            };
-            let json = serde_json::to_string_pretty(&result)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(&result).ok();
-            tool_response_named(&db, json, structured, "cartog_trace", Some(&from), stale)
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Show inheritance hierarchy for a class.
-    #[tool(
-        description = "Show class inheritance tree. Use when asked 'show the class hierarchy', 'what extends X?', 'what does X inherit from?'. Not for: trait/interface implementations (use cartog_refs with kind=implements). Returns: array of {child: string, parent: string} (symbol names) ordered top-down.",
-        annotations(
-            title = "Class hierarchy",
-            read_only_hint = true,
-            open_world_hint = false
-        ),
-        output_schema = output_schema_for::<HierarchyList>()
-    )]
-    async fn cartog_hierarchy(
-        &self,
-        Parameters(params): Parameters<HierarchyParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let name = params.name;
-        let db = Arc::clone(&self.db);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            debug!(name = %name, "hierarchy");
-            let db = db.lock().map_err(|_| {
-                mcp_err("internal error: database lock poisoned (server restart required)")
-            })?;
-            let pairs = db
-                .hierarchy(&name)
-                .map_err(|e| mcp_err(format!("hierarchy query failed: {e}")))?;
-
-            let entries: Vec<HierarchyEntry> = pairs
-                .into_iter()
-                .map(|(child, parent)| HierarchyEntry { child, parent })
-                .collect();
-
-            let json = serde_json::to_string_pretty(&entries)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(HierarchyList { results: entries }).ok();
-            tool_response_named(
-                &db,
-                json,
-                structured,
-                "cartog_hierarchy",
-                Some(&name),
-                stale,
-            )
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// File-level import dependencies.
-    #[tool(
-        description = "Show what a file imports. Use when asked 'what does this file depend on?', 'show imports for X'. Not for: reverse dependencies (use cartog_refs with kind=imports on the imported module). Returns: Edge[] of {target_name, line} per import statement.",
-        annotations(
-            title = "File dependencies",
-            read_only_hint = true,
-            open_world_hint = false
-        ),
-        output_schema = output_schema_for::<EdgeList>()
-    )]
-    async fn cartog_deps(
-        &self,
-        Parameters(params): Parameters<DepsParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let file = params.file;
-        let db = Arc::clone(&self.db);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            debug!(file = %file, "deps");
-            let db = db.lock().map_err(|_| {
-                mcp_err("internal error: database lock poisoned (server restart required)")
-            })?;
-            let edges = db
-                .file_deps(&file)
-                .map_err(|e| mcp_err(format!("deps query failed: {e}")))?;
-
-            let json = serde_json::to_string_pretty(&edges)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(EdgeList { results: edges }).ok();
-            tool_response(&db, json, structured, "cartog_deps", stale)
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Search for symbols by name — use this to discover exact names before calling refs/callees/impact.
-    #[tool(
-        description = "Find symbols by exact or partial name. Use ONLY to get a precise symbol name before calling cartog_refs, cartog_callees, or cartog_impact. Not for: general code discovery (use cartog_rag_search instead — better recall for natural-language queries). Supports prefix and substring matching, case-insensitive. Returns: Symbol[] ranked by centrality (most-referenced first).",
-        annotations(
-            title = "Search symbols",
-            read_only_hint = true,
-            open_world_hint = false
-        ),
-        output_schema = output_schema_for::<SymbolList>()
-    )]
-    async fn cartog_search(
-        &self,
-        Parameters(params): Parameters<SearchParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let query = params.query;
-        let kind_str = params.kind;
-        let file = params.file;
-        let limit = params.limit.unwrap_or(30).min(MAX_SEARCH_LIMIT);
-        let db = Arc::clone(&self.db);
-        let cwd = Arc::clone(&self.cwd);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            if query.is_empty() {
-                return Err(mcp_err("query cannot be empty"));
-            }
-
-            let kind_filter = kind_str
-                .as_deref()
-                .map(|s| {
-                    s.parse::<cartog_core::SymbolKind>().map_err(|_| {
-                        mcp_err(
-                            "invalid symbol kind. Valid: function, class, method, variable, import, interface, enum, type-alias, trait, module, document",
-                        )
-                    })
-                })
-                .transpose()?;
-
-            // Validate file path is within CWD — consistent with cartog_outline / cartog_deps.
-            let validated_file: Option<String> = file
-                .map(|f| {
-                    validate_path_within_cwd_canonical(&f, &cwd)
-                        .map_err(mcp_err)
-                        .map(|p| p.to_string_lossy().into_owned())
-                })
-                .transpose()?;
-            let file_filter = validated_file.as_deref();
-            debug!(query = %query, kind = ?kind_filter, limit, "search");
-            let db = db.lock().map_err(|_| mcp_err("internal error: database lock poisoned (server restart required)"))?;
-            let mut symbols = db
-                .search(&query, kind_filter, file_filter, limit)
-                .map_err(|e| mcp_err(format!("search failed: {e}")))?;
-            if mcp_compact() {
-                symbols.compact_in_place();
-            }
-
-            let json = serde_json::to_string_pretty(&symbols)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(SymbolList { results: symbols }).ok();
-            tool_response(&db, json, structured, "cartog_search", stale)
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Index statistics summary.
-    #[tool(
-        description = "Show index health: file count, symbol count, edge count, and edge resolution buckets. Use to verify the index is built and check coverage. Not for: finding code (use cartog_search or cartog_rag_search). Returns: {num_files, num_symbols, num_edges, num_resolved, num_unresolvable, num_external, languages, symbol_kinds, role, watcher_active}. num_external counts edges whose LSP-resolved target lives outside the indexed root (stdlib, deps, node_modules).",
-        annotations(title = "Index stats", read_only_hint = true, open_world_hint = false),
-        output_schema = output_schema_for::<StatsResult>()
-    )]
-    async fn cartog_stats(&self) -> Result<CallToolResult, McpError> {
-        let db = Arc::clone(&self.db);
-        let role = self.role.load();
-        let watcher_active = self
-            .watcher_active
-            .load(std::sync::atomic::Ordering::Relaxed);
-
-        tokio::task::spawn_blocking(move || {
-            debug!("stats");
-            let db = db.lock().map_err(|_| {
-                mcp_err("internal error: database lock poisoned (server restart required)")
-            })?;
-            let stats = db
-                .stats()
-                .map_err(|e| mcp_err(format!("stats query failed: {e}")))?;
-
-            // `watcher_active=false` on a Primary means either the user did not
-            // request `--watch`, or a post-promotion watcher spawn failed (e.g.,
-            // another live `cartog watch` holds the watch slot, or notify install
-            // failed). The user can distinguish the cases by whether they passed
-            // `--watch`.
-            let result = StatsResult {
-                stats,
-                role,
-                watcher_active,
-            };
-            let json = serde_json::to_string_pretty(&result)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(&result).ok();
-            // cartog_stats bypasses tool_response_named so it must log itself
-            // — otherwise MCP-side stats calls disappear from
-            // `cartog stats --savings`.
-            log_tool_query(&db, "cartog_stats");
-            Ok(success_result(json, structured))
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Arm a deferred cartog self-update.
-    ///
-    /// Deliberately NOT gated by `refuse_if_read_only`: arming writes the
-    /// machine-level state file, not the index DB, so a read-only secondary
-    /// may arm just like the primary. The guard exists only to stop a
-    /// secondary from writing the DB — adding it here would be a category
-    /// error.
-    ///
-    /// Never swaps the binary in-session: this server IS the live peer that
-    /// `cartog self update` refuses to overwrite. It shells out to
-    /// `self update --defer`, which records the target and exits without
-    /// touching the binary; the boundary swap happens at SessionEnd.
-    #[tool(
-        description = "Arm a deferred cartog self-update. Does NOT upgrade in this session — the running server keeps its current binary; the new version becomes active after this session ends (or the next restart). Use when the user confirms they want to update cartog. When cartog is installed as a Claude Code plugin, this arms the plugin's PINNED version (discovered from the plugin manifest); otherwise it arms the latest stable release. Not for: indexing or search. Returns: {current, target, status, apply, message}.",
-        annotations(
-            title = "Update cartog",
-            read_only_hint = false,
-            destructive_hint = false,
-            // Not idempotent: a latest-release arm re-fetches the tag and each
-            // arm rewrites armed_at / last_update_check timestamps, so repeated
-            // calls change state.toml even when the armed target is unchanged.
-            idempotent_hint = false,
-            open_world_hint = true
-        ),
-        output_schema = output_schema_for::<UpdateResult>()
-    )]
-    async fn cartog_update(&self) -> Result<CallToolResult, McpError> {
-        tokio::task::spawn_blocking(move || {
-            debug!("update (arm deferred)");
-            let exe = std::env::current_exe()
-                .map_err(|e| mcp_err(format!("cannot resolve cartog binary: {e}")))?;
-            // Arm the plugin's pinned version when discoverable so we can't
-            // overshoot the pin; fall back to latest stable otherwise.
-            let mut args = vec!["self", "update", "--defer", "--json"];
-            let pin = discover_plugin_pin();
-            if let Some(ref v) = pin {
-                args.push("--to");
-                args.push(v);
-            }
-            let output = std::process::Command::new(exe)
-                .args(&args)
-                .output()
-                .map_err(|e| mcp_err(format!("failed to run self update --defer: {e}")))?;
-
-            let result = parse_arm_output(&output);
-            let json = serde_json::to_string_pretty(&result)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(&result).ok();
-            Ok(success_result(json, structured))
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Codebase orientation: file list + top symbols by centrality.
-    #[tool(
-        description = "Orient yourself in an unfamiliar codebase. Returns the full file list plus the top N symbols ranked by reference count (most-used definitions first). Use as the FIRST call when dropped into a new repo, before search or refs. Not for: locating a specific symbol (use cartog_search), or fetching one file's structure (use cartog_outline). Returns: {files: string[], top_symbols: Symbol[]}.",
-        annotations(title = "Codebase map", read_only_hint = true, open_world_hint = false),
-        output_schema = output_schema_for::<MapResult>()
-    )]
-    async fn cartog_map(
-        &self,
-        Parameters(params): Parameters<MapParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let limit = params.limit.unwrap_or(50);
-        let db = Arc::clone(&self.db);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            debug!(limit, "map");
-            let db = db.lock().map_err(|_| {
-                mcp_err("internal error: database lock poisoned (server restart required)")
-            })?;
-            let files = db
-                .all_files()
-                .map_err(|e| mcp_err(format!("files query failed: {e}")))?;
-            let mut top_symbols = db
-                .top_symbols(limit)
-                .map_err(|e| mcp_err(format!("top_symbols query failed: {e}")))?;
-            if mcp_compact() {
-                top_symbols.compact_in_place();
-            }
-
-            let result = MapResult { files, top_symbols };
-
-            let json = serde_json::to_string_pretty(&result)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(&result).ok();
-            tool_response(&db, json, structured, "cartog_map", stale)
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Show symbols affected by recent git changes.
-    #[tool(
-        description = "Show what changed recently. Symbols affected by the last N git commits plus working-tree changes. Use when asked 'what changed?', 'what did I modify?', or to understand recent code activity before a review. Not for: arbitrary git diffs (use Bash with `git diff`). Returns: {changed_files: string[], symbols: Symbol[]}.",
-        annotations(
-            title = "Recent changes",
-            read_only_hint = true,
-            open_world_hint = false
-        ),
-        output_schema = output_schema_for::<cartog_core::ChangesResult>()
-    )]
-    async fn cartog_changes(
-        &self,
-        Parameters(params): Parameters<ChangesParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let commits = params.commits.unwrap_or(5);
-        let kind_str = params.kind;
-        let db = Arc::clone(&self.db);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            let kind_filter = kind_str
-                .as_deref()
-                .map(|s| {
-                    s.parse::<cartog_core::SymbolKind>().map_err(|_| {
-                        mcp_err(
-                            "invalid symbol kind. Valid: function, class, method, variable, import, interface, enum, type-alias, trait, module, document",
-                        )
-                    })
-                })
-                .transpose()?;
-
-            debug!(commits, kind = ?kind_filter, "changes");
-
-            let root = std::env::current_dir()
-                .map_err(|e| mcp_err(format!("cannot determine CWD: {e}")))?;
-
-            let changed_files = indexer::git_recently_changed_files(&root, commits)
-                .map_err(|e| mcp_err(format!("git changes failed: {e}")))?;
-
-            let db = db.lock().map_err(|_| mcp_err("internal error: database lock poisoned (server restart required)"))?;
-            let mut symbols = db
-                .symbols_for_files(&changed_files, kind_filter)
-                .map_err(|e| mcp_err(format!("symbols query failed: {e}")))?;
-            if mcp_compact() {
-                symbols.compact_in_place();
-            }
-
-            let result = cartog_core::ChangesResult {
-                changed_files,
-                symbols,
-            };
-
-            let json = serde_json::to_string_pretty(&result)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(&result).ok();
-            tool_response(&db, json, structured, "cartog_changes", stale)
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Build embedding index for semantic code search.
-    #[tool(
-        description = "Build the embedding index for semantic search. Optional — cartog_rag_search ALREADY works at FTS5 (BM25) quality without embeddings; only run this when you want vector recall on top. Requires `cartog rag setup` from the CLI first to download the model. Not for: first-time setup of cartog (cartog_index is what you want). Returns: {embedded, skipped, failed, dim}.",
-        annotations(
-            title = "Build embedding index",
-            read_only_hint = false,
-            destructive_hint = false,
-            idempotent_hint = false,
-            open_world_hint = false
-        )
-    )]
-    async fn cartog_rag_index(
-        &self,
-        Parameters(params): Parameters<RagIndexParams>,
-        ctx: RequestContext<RoleServer>,
-    ) -> Result<CallToolResult, McpError> {
-        if let Some(err) = self.refuse_if_read_only("cartog_rag_index") {
-            return Err(err);
-        }
-        let path = params.path;
-        let force = params.force;
-        let db = Arc::clone(&self.db);
-        let cwd = Arc::clone(&self.cwd);
-        let redact = self.redact;
-        let walk_filter = Arc::clone(&self.walk_filter);
-        let provider = Arc::clone(&self.embedding_provider);
-
-        let (progress_tx, forwarder) = match ctx.meta.get_progress_token() {
-            Some(token) => {
-                let notifier = progress::peer_notifier(ctx.peer.clone());
-                let fwd = progress::spawn_forwarder(token, notifier);
-                (Some(fwd.tx.clone()), Some(fwd))
-            }
-            None => (None, None),
-        };
-
-        let cancel = ctx.ct.clone();
-        let join = tokio::task::spawn_blocking(move || {
-            let validated = validate_path_within_cwd_canonical(&path, &cwd).map_err(mcp_err)?;
-            debug!(path = %validated.display(), force, "rag index");
-
-            let db = db.lock().map_err(|_| {
-                mcp_err("internal error: database lock poisoned (server restart required)")
-            })?;
-
-            let probe = || cancel.is_cancelled();
-            let probe_ref: Option<&(dyn Fn() -> bool + Send + Sync)> = Some(&probe);
-
-            // Ensure the code graph index is up to date first. Inner phases are
-            // intentionally suppressed: the RAG tool exposes only rag-specific
-            // phases (preparing/embedding/storing) so the client-facing
-            // vocabulary stays stable.
-            let _ = indexer::index_directory(
-                &db,
-                &validated,
-                false,
-                false,
-                None,
-                probe_ref,
-                redact,
-                &std::collections::HashMap::new(),
-                walk_filter.as_ref(),
-            )
-            .map_err(|e| mcp_err(format!("code graph indexing failed: {e}")))?;
-
-            let mut provider = provider.lock().map_err(|_| {
-                mcp_err(
-                    "internal error: embedding provider lock poisoned (server restart required)",
-                )
-            })?;
-
-            let rag_cb = progress_tx
-                .as_ref()
-                .map(|tx| progress::rag_callback(tx.clone()));
-            let cb_ref: Option<&(dyn Fn(rag::indexer::ProgressUpdate) + Send + Sync)> =
-                rag_cb.as_ref().map(|f| f as _);
-
-            let result =
-                rag::indexer::index_embeddings(&db, provider.as_mut(), force, cb_ref, probe_ref)
-                    .map_err(|e| mcp_err(format!("embedding indexing failed: {e}")))?;
-
-            let json = serde_json::to_string_pretty(&result)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            Ok(CallToolResult::success(vec![Content::text(json)]))
-        })
-        .await;
-
-        // Drain the forwarder unconditionally — see cartog_index for rationale.
-        if let Some(fwd) = forwarder {
-            drop(fwd.tx);
-            let _ = fwd.join.await;
-        }
-
-        join.map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Semantic search over code symbols using hybrid FTS5 + vector search.
-    #[tool(
-        description = "Search code by concept, keyword, or natural language. Returns ranked symbols with snippet excerpts — locations + previews, not full bodies. Use to LOCATE code matching a concept: 'find code related to...', 'show me the authentication logic'. For 'how does X work?' or understanding an area, prefer cartog_context — it returns full bodies + call neighbors in one call instead of snippets you'd then have to Read. Works even without embeddings (keyword matching alone is already strong). Prefer this over Grep for code discovery. Not for: looking up a known symbol name (use cartog_search instead — more precise). Filter with kind='document' for docs, kind='all' for both. Returns: Symbol[] ranked by relevance with snippet excerpts.",
-        annotations(
-            title = "Semantic code search",
-            read_only_hint = true,
-            open_world_hint = false
-        ),
-        output_schema = output_schema_for::<rag::search::HybridSearchResult>()
-    )]
-    async fn cartog_rag_search(
-        &self,
-        Parameters(params): Parameters<RagSearchParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let query = params.query;
-        let kind_str = params.kind;
-        let limit = params.limit.unwrap_or(10).min(MAX_SEARCH_LIMIT);
-        let db = Arc::clone(&self.db);
-        let provider = Arc::clone(&self.embedding_provider);
-        let reranker = Arc::clone(&self.reranker_provider);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            if query.is_empty() {
-                return Err(mcp_err("query cannot be empty"));
-            }
-
-            debug!(query = %query, kind = ?kind_str, limit, "rag search");
-            let db = db.lock().map_err(|_| mcp_err("internal error: database lock poisoned (server restart required)"))?;
-
-            let kind_filter = match kind_str.as_deref() {
-                Some("all") => rag::search::KindFilter::All,
-                Some(s) => {
-                    let kind = s.parse::<cartog_core::SymbolKind>().map_err(|_| {
-                        mcp_err(
-                            "invalid symbol kind. Valid: function, class, method, variable, import, interface, enum, type-alias, trait, module, document, all",
-                        )
-                    })?;
-                    rag::search::KindFilter::Exact(kind)
-                }
-                None => rag::search::KindFilter::CodeOnly,
-            };
-
-            let mut provider = provider
-                .lock()
-                .map_err(|_| mcp_err("internal error: embedding provider lock poisoned (server restart required)"))?;
-            let mut reranker = reranker
-                .lock()
-                .map_err(|_| mcp_err("internal error: reranker lock poisoned (server restart required)"))?;
-            let mut result = match reranker.as_mut() {
-                Some(r) => rag::search::hybrid_search(
-                    &db, &query, limit, kind_filter, provider.as_mut(), Some(r.as_mut()),
-                ),
-                None => rag::search::hybrid_search(
-                    &db, &query, limit, kind_filter, provider.as_mut(), None,
-                ),
-            }.map_err(|e| mcp_err(format!("semantic search failed: {e}")))?;
-            // Compact-by-default: bound each body to a snippet (the tool advertises
-            // "snippet excerpts"). CARTOG_MCP_COMPACT=0 restores full bodies.
-            if mcp_compact() {
-                result.snippet_in_place();
-            }
-
-            let json = serde_json::to_string_pretty(&result)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(&result).ok();
-            tool_response(&db, json, structured, "cartog_rag_search", stale)
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
-    }
-
-    /// Build a one-shot task-context bundle fusing semantic search, structural
-    /// neighbors, and centrality.
-    #[tool(
-        description = "PRIMARY TOOL — call FIRST for any 'how does X work?', 'understand/survey area Y', 'where do I implement Z?' question. ONE call returns the relevant symbols (semantic + keyword), their 1-hop call neighbors, and high-centrality definitions in the same files, with bodies inline and budgeted to fit. Read-equivalent: do NOT re-open the symbols it returns, and usually the ONLY call you need — answer from its bundle instead of a chain of search/refs/callees/outline/Read. Drill in with the granular tools (cartog_refs, cartog_callees, cartog_trace) only for follow-ups it doesn't cover. Routes through semantic search, so it finds code by concept, not just name. Raise `tokens` (default 6000, max 20000) for a whole subsystem. Not for: a single known symbol (use cartog_search), or one specific call path (use cartog_trace). Returns: {task, entries: [{symbol, reason, score, body?}], approx_tokens}.",
-        annotations(
-            title = "Task context bundle",
-            read_only_hint = true,
-            open_world_hint = false
-        ),
-        output_schema = output_schema_for::<rag::context::TaskContext>()
-    )]
-    async fn cartog_context(
-        &self,
-        Parameters(params): Parameters<ContextParams>,
-    ) -> Result<CallToolResult, McpError> {
-        let task = params.task;
-        let tokens = params
-            .tokens
-            .unwrap_or(DEFAULT_CONTEXT_TOKENS)
-            .min(MAX_CONTEXT_TOKENS);
-        let db = Arc::clone(&self.db);
-        let provider = Arc::clone(&self.embedding_provider);
-        let reranker = Arc::clone(&self.reranker_provider);
-        let stale = self.stale_snapshot();
-
-        tokio::task::spawn_blocking(move || {
-            if task.is_empty() {
-                return Err(mcp_err("task description cannot be empty"));
-            }
-            debug!(task = %task, tokens, "context");
-            let db = db.lock().map_err(|_| {
-                mcp_err("internal error: database lock poisoned (server restart required)")
-            })?;
-            let mut provider = provider.lock().map_err(|_| {
-                mcp_err(
-                    "internal error: embedding provider lock poisoned (server restart required)",
-                )
-            })?;
-            let mut reranker = reranker.lock().map_err(|_| {
-                mcp_err("internal error: reranker lock poisoned (server restart required)")
-            })?;
-            let opts = rag::context::ContextOptions::default();
-            let result = match reranker.as_mut() {
-                Some(r) => rag::context::build_task_context(
-                    &db,
-                    &task,
-                    tokens,
-                    provider.as_mut(),
-                    Some(r.as_mut()),
-                    &opts,
-                ),
-                None => rag::context::build_task_context(
-                    &db,
-                    &task,
-                    tokens,
-                    provider.as_mut(),
-                    None,
-                    &opts,
-                ),
-            }
-            .map_err(|e| mcp_err(format!("context build failed: {e}")))?;
-            // Compact trims per-entry symbol noise but KEEPS the budgeted bodies —
-            // this tool's whole value is its inline bodies.
-            let mut result = result;
-            if mcp_compact() {
-                result.compact_in_place();
-            }
-
-            let json = serde_json::to_string_pretty(&result)
-                .map_err(|e| mcp_err(format!("serialization failed: {e}")))?;
-            let structured = serde_json::to_value(&result).ok();
-            tool_response(&db, json, structured, "cartog_context", stale)
-        })
-        .await
-        .map_err(|e| mcp_err(format!("task join failed: {e}")))?
+    /// Combined router over every per-domain tool block. Hand-written so the
+    /// `tool_router` name the `#[tool_handler]` default and tests expect still
+    /// resolves; each `*_router()` is generated by the `#[tool_router(router=...)]`
+    /// on its tools/ submodule impl block.
+    fn tool_router() -> rmcp::handler::server::router::tool::ToolRouter<Self> {
+        Self::index_router()
+            + Self::graph_router()
+            + Self::search_router()
+            + Self::rag_router()
+            + Self::manage_router()
     }
 }
