+++ crates/cartog-mcp/src/tools/mod.rs
@@ -1,0 +1,8 @@
+//! MCP tool implementations, grouped by concern. Each submodule adds a
+//! `#[tool_router(router = ...)]` impl block; lib.rs combines them.
+
+mod graph;
+mod index;
+mod manage;
+mod rag;
+mod search;
